some refactor + started implementing birthday predicate
