+++ src/main/kotlin/day1/homeExercise/Domain.kt
@@ -1,0 +1,12 @@
+package day1.homeExercise
+
+import java.time.LocalDate
+
+data class Employee(
+    val lastName: String,
+    val firstName: String,
+    val birthDate: LocalDate,
+    val email: Email
+)
+
+data class Email(val value: String)
